+++ frontend/src/utils/cookies.js
@@ -1,43 +1,57 @@
 export default class Cookies {
   static get(name) {
-    if (document.cookie.length === 0) return null;
-    var c_start = document.cookie.indexOf(`${name}=`);
-    if (c_start === -1) return null;
-    c_start = c_start + name.length + 1;
-    var c_end = document.cookie.indexOf(";", c_start);
-    if (c_end == -1) c_end = document.cookie.length;
-    return decodeURIComponent(document.cookie.substring(c_start, c_end));
+    if (document.cookie.length === 0) {
+      return null;
+    }
+
+    const start = document.cookie.indexOf(`${name}=`);
+    if (start == -1) {
+      return null;
+    }
+
+    const end = document.cookie.indexOf(";", start);
+    const prefix = start + name.length + 1;
+    const suffix = end == -1 ? document.cookie.length : "";
+    const rawcookie = document.cookie.substring(prefix, suffix);
+
+    return decodeURIComponent(rawcookie);
   }
 
-  static set(name, value, hours) {
-    if (hours > 0) {
-      let now = new Date();
-      now.setTime(now.getTime() + hours * 3600 * 1000);
-      let date = now.toUTCString();
-      document.cookie =
-        name + `=${encodeURIComponent(value)}; expires=${date}; path=/`;
-    } else {
+  static set(name, value, hours = 0) {
+    if (!name || !value) {
+      console.error("name or value param missing, name or value will be empty in cookie : got :", { name, param });
+    }
+
+    if (hours <= 0) {
       document.cookie = name + `=${encodeURIComponent(value)}; path=/`;
+      return;
+    }
+    
+    const now = new Date();
+    now.setTime(now.getTime() + hours * 3600 * 1000);
+    document.cookie = `${name}=${encodeURIComponent(value)}; expires=${now.toUTCString()}; path=/`;
+  }
+
+  static remove(name = undefined) {
+    if (name) {
+      document.cookie = name + `=''; expires=${new Date(1).toUTCString()}`;
     }
   }
 
-  static remove(name) {
-    if (name)
-      document.cookie = name + `=''; expires=${new Date(1).toUTCString()}`;
-  }
-
   static getAll() {
-    if (document.cookie.length === 0) return null;
-    var cookies = {};
-    document.cookie.split(";").forEach((pairs) => {
-      let pair = pairs.split("=");
+    if (document.cookie.length === 0) {
+      return null;
+    }
+
+    return document.cookie.split(";").reduce((cookies, pairs) => {
+      const pair = pairs.split("=");
       cookies[(pair[0] + "").trim()] = decodeURIComponent(pair[1]);
-    });
-    return cookies;
+      return cookies;
+    }, {});
   }
 
-  static check(name) {
-    name = this.get(name);
-    return name && name !== "" ? true : false;
+  static check(name = undefined) {
+    const n = this.get(name);
+    return n && n !== "" ? true : false;
   }
 }