+++ frontend/CHANGELOG.md
@@ -36,7 +36,9 @@
 
 ##### New Features
 
-* **frontend:**  add generate changelog (9ccd9b5d)
+* **frontend:**
+  *  begin adding trade planner form (7b8291e0)
+  *  add generate changelog (9ccd9b5d)
 *  add generate-changelog for frontend on each commit (08678a5d)
 *  adding pages to routes (a97e3d5c)
 *  continue adding router/routes/pages (a5bb555c)